Finished bound objects functionality

+++ st_slipspace/include/gr_tour_object.h
@@ -11,7 +11,7 @@ public:
     };
     virtual ~grTourObject(){};
     virtual void startup(gfArchive* archive, u32 unk1, gfSceneRoot::LayerType layerType);
-    virtual void update(int deltaFrame);
+    virtual void update(float deltaFrame);
 
     static grTourObject* create(int mdlIndex, const char* tgtNodeName, const char* taskName);
