+++ st_slipspace/source/gr_tour_object.cpp
@@ -20,7 +20,7 @@ void grTourObject::startup(gfArchive* archive, u32 unk1, gfSceneRoot::LayerType 
     grMadein::startup(archive, unk1, layerType);
 }
 
-void grTourObject::update(int deltaFrame)
+void grTourObject::update(float deltaFrame)
 {
     this->setEnableCollisionStatus(true);
 }
